Cambio decimales a enteros Tran. (DB reset)

+++ core/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class PerfilUsuario(models.Model):
     usuario = models.OneToOneField(User, on_delete=models.CASCADE)
-    saldo = models.DecimalField(max_digits=10, decimal_places=2, default=10000)
+    saldo = models.IntegerField(default=10000)
 
     def __str__(self):
         return self.usuario.username
@@ -11,7 +11,7 @@ class PerfilUsuario(models.Model):
 class Transferencia(models.Model):
     destinatario = models.ForeignKey(PerfilUsuario, on_delete=models.CASCADE, related_name='transferencias_recibidas')
     remitente = models.ForeignKey(PerfilUsuario, on_delete=models.CASCADE, related_name='transferencias_enviadas')
-    monto = models.DecimalField(max_digits=10, decimal_places=2)
+    monto = models.IntegerField()
     fecha = models.DateTimeField(auto_now_add=True)
 
     def __str__(self):
